A better Day 1 implementation.
After working on a Haskell implementation of this I came across a much
cleaner approach that also works well in Scala. The idea is to create
an infinite Stream that cycles through the input data and then zip
that stream with the original data after dropping some number of
values from the start of the stream in order to line up the values to
compare and sum.